refactor(canvas-viewer): dedupe the font-loading test fakes and tidy comments
Extract the FakeFontFace/installFakeFontApis pair duplicated across
font-loading.test.tsx and use-viewer-font-ready.test.tsx into
test-utils/fake-font-face.ts, dropping the `any` casts on the globalThis
writes along the way. Replace the definite-assignment `timeoutHandle!`
with an optional handle, rewrite the comments that carried task/PR
chronology as enduring rationale, and let Biome organize the imports the
previous commit left unsorted.

Behavior unchanged.
