+++ packages/canvas-viewer/src/index.ts
@@ -8,9 +8,9 @@ export { CanvasViewer } from './CanvasViewer.js'
 export { VIEWER_FONT_FAMILY } from './font.js'
 export { ensureViewerFontLoaded, type ViewerFontStatus } from './font-loading.js'
 export { createBrowserMeasureText } from './measure-text.js'
-export { useViewerFontReady } from './use-viewer-font-ready.js'
 export type { CanvasViewerHandle, MountCanvasViewerOptions } from './mount.js'
 export { mountCanvasViewer, ViewerSceneError } from './mount.js'
 export type { ViewerScene } from './scene.js'
 export { parseViewerScene, serializeViewerScene, viewerSceneSchema } from './scene.js'
+export { useViewerFontReady } from './use-viewer-font-ready.js'
 export { serializeSceneForScriptTag } from './widget/embed-scene.js'
